Drop useless casts in dynamic core decoder

+++ src/cpu/core_dyn_x86/decoder.h
@@ -2900,7 +2900,7 @@ finish_block:
 	/* Setup the correct end-address */
 	decode.active_block->page.end=--decode.page.index;
 	dyn_mem_execute(cache_addr, cache_bytes);
-	const auto cache_flush_bytes = static_cast<size_t>(decode.block->cache.size);
+	const auto cache_flush_bytes = decode.block->cache.size;
 	dyn_cache_invalidate(cache_addr, cache_flush_bytes);
 	assert(decode.block->cache.size <= cache_bytes);
 	//	LOG_MSG("Created block size %d start %d end
